refactor: remove unnecessary private mod shared (#400)
* refactor: remove unnecessary private mod shared

* chore: make fmt happy
