+++ .changes/remove-shared-mod.md
@@ -1,0 +1,6 @@
+---
+"wry": patch
+---
+
+The only thing that private mod shared does is re-export http mod to public, 
+we can just pub mod http.
