+++ src/lib.rs
@@ -89,7 +89,7 @@ use std::sync::mpsc::{RecvError, SendError};
 
 use crate::{
   application::window::BadIcon,
-  shared::http::{
+  http::{
     header::{InvalidHeaderName, InvalidHeaderValue},
     method::InvalidMethod,
     status::InvalidStatusCode,
@@ -100,10 +100,8 @@ pub use serde_json::Value;
 use url::ParseError;
 
 pub mod application;
+pub mod http;
 pub mod webview;
-
-mod shared;
-pub use shared::*;
 
 /// Convenient type alias of Result type for wry.
 pub type Result<T> = std::result::Result<T, Error>;
