--- src/shared/mod.rs
@@ -1,6 +1,0 @@
-// Copyright 2019-2021 Tauri Programme within The Commons Conservancy
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-License-Identifier: MIT
-
-/// HTTP types used by wry protocol.
-pub mod http;
